+++ client/src/components/Live.vue
@@ -27,6 +27,7 @@ export default defineComponent({
   left: 0;
   bottom: 0;
   right: 0;
+  pointer-events: all;
 }
 .iframe {
   height: 100%;
